Working HEX transmissions.
Some caveats

If any null bytes 0x00 are sent, the modem returns Error -23.
If too many bytes are sent (about 50) the modem returns Error -21.
If the modem is asked to transmit more than DR0, then the modem returns Error -3.

So many messages.

And still unsure how to format the Serial.print() to ensure that it always produces exactly two hex nibbles for each byte. I have an unfounded feeling that 0x01 is being printed as a single 1 character, rather than 01.